Moved most of Depth_Process to dbsgui/process/trajectory.py Added some debug print statements.
